requirements.txt
nltk==3.5
pandas==1.3.3
Django==3.2.5
numpy==1.20.3
tweepy==4.0.0
djangorestframework==3.13.1
python_dateutil==2.8.2
